api done

+++ internal/app/room/usecase/roomUsecase.go
@@ -1,8 +1,12 @@
 package roomUsecase
 
 import (
+	"errors"
+
 	"github.com/Kostikans/avitoTest/internal/app/room"
 	roomModel "github.com/Kostikans/avitoTest/internal/app/room/model"
+	"github.com/Kostikans/avitoTest/internal/package/clientError"
+	"github.com/Kostikans/avitoTest/internal/package/customError"
 )
 
 type RoomUsecase struct {
@@ -18,9 +22,20 @@ func (rUsecase *RoomUsecase) AddRoom(room roomModel.RoomAdd) (roomModel.RoomID, 
 }
 
 func (rUsecase *RoomUsecase) DeleteRoom(roomID int64) error {
+	exist, err := rUsecase.CheckRoomExist(roomID)
+	if err != nil {
+		return err
+	}
+	if exist == false {
+		return customError.NewCustomError(errors.New("room doesn't exist"), clientError.NotFound, 1)
+	}
 	return rUsecase.RoomRepo.DeleteRoom(roomID)
 }
 
 func (rUsecase *RoomUsecase) GetRooms(order roomModel.RoomOrder) ([]roomModel.Room, error) {
 	return rUsecase.RoomRepo.GetRooms(order)
+}
+
+func (rUsecase *RoomUsecase) CheckRoomExist(roomID int64) (bool, error) {
+	return rUsecase.RoomRepo.CheckRoomExist(roomID)
 }
